trade2/requirements.txt
ccxt>=4.1.59
python-telegram-bot>=20.3
python-dotenv>=1.0.0
numpy==1.24.3
pandas==2.0.3
python-binance==1.0.19
python-telegram-bot==20.3
matplotlib==3.5.1
aiohttp==3.8.1
sqlalchemy==2.0.0
